requirements.txt
Django==3.2.5
pandas==1.3.1
psutil==5.8.0
Unidecode==1.2.0
celery_progress==0.1.1
tqdm==4.62.0
django_rest_knox==4.1.0
holidays==0.11.2
djangorestframework==3.12.4
knox==0.1.14
pyexcel_xlsx==0.6.0
celery~=5.1.2
sqlalchemy~=1.4.23